backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pyyaml>=6.0
python-multipart>=0.0.20
watchdog>=6.0.0
pydantic>=2.11.0
